Lift compute_msbar_masses to Runner and relative tests
