+++ src/eko/msbar_masses.py
@@ -198,3 +198,62 @@ def evolve_msbar_mass(
     else:
         ev_mass = msbar_ker_dispatcher(q2_to, q2m_ref, strong_coupling, fact_to_ren, nf)
         return m2_ref * ev_mass ** 2
+
+
+def compute_msbar_mass(theory_card):
+    r"""
+    Compute the :math:`\overline{MS}` masses solving the equation :math:`m_{\bar{MS}}(m) = m`
+
+    Parameters
+    ----------
+        theory_card: dict
+            theory run card
+
+    Returns
+    -------
+        msbar_masses: list
+            list of msbar masses
+    """
+    heavy_flavors = "cbt"
+    msbar_masses = np.full(3, np.inf)
+    shift = 3
+    q2m = np.power([theory_card[f"Qm{q}"] for q in heavy_flavors], 2)
+    m2 = np.power([theory_card[f"m{q}"] for q in heavy_flavors], 2)
+    config = {
+        "q2m_ref": np.power([theory_card[f"Qm{q}"] for q in heavy_flavors], 2),
+        "as_ref": theory_card["alphas"],
+        "q2a_ref": np.power(theory_card["Qref"], 2),
+        "order": theory_card["PTO"],
+        "fact_to_ren": theory_card["fact_to_ren_scale_ratio"] ** 2,
+    }
+    for nf in [3, 4, 5]:
+        q2m_ref = q2m[nf - shift]
+        m2_ref = m2[nf - shift]
+        # check if mass is already given at the pole
+        if q2m_ref == m2_ref:
+            msbar_masses[nf - shift] = m2_ref
+            continue
+        # if self.q2_ref > q2m_ref:
+        #     raise ValueError("In MSBAR scheme Q0 must be lower than any Qm")
+        # if q2m_ref > m2_ref:
+        #     raise ValueError("In MSBAR scheme each heavy quark \
+        #         mass reference scale must be smaller or equal than \
+        #             the value of the mass itself"
+        #     )
+        config["thr_masses"] = msbar_masses
+        msbar_masses[nf - shift] = evolve_msbar_mass(m2_ref, q2m_ref, nf, config=config)
+
+    # Check the msbar ordering
+    for nf in [4, 5]:
+        q2m_ref = q2m[nf - shift]
+        m2_ref = m2[nf - shift]
+        m2_msbar = msbar_masses[nf - shift - 1]
+        config["thr_masses"] = msbar_masses
+        # check that m_msbar_hq < msbar_hq+1 (m_msbar_hq)
+        m2_test = evolve_msbar_mass(m2_ref, q2m_ref, nf, config=config, q2_to=m2_msbar)
+        if m2_msbar > m2_test:
+            raise ValueError(
+                "The MSBAR masses do not preserve the correct ordering,\
+                     check the inital reference values"
+            )
+    return msbar_masses
