+++ tests/test_ome.py
@@ -98,7 +98,7 @@ def test_quad_ker(monkeypatch):
             backward_method=None,
             a_s=0.0,
             L=0.0,
-            is_msbar=False
+            is_msbar=False,
         )
         np.testing.assert_allclose(res_ns, 1.0)
         res_s = quad_ker(
@@ -111,7 +111,7 @@ def test_quad_ker(monkeypatch):
             backward_method=None,
             a_s=0.0,
             L=0.0,
-            is_msbar=False
+            is_msbar=False,
         )
         np.testing.assert_allclose(res_s, 1.0)
         res_s = quad_ker(
@@ -124,7 +124,7 @@ def test_quad_ker(monkeypatch):
             backward_method=None,
             a_s=0.0,
             L=0.0,
-            is_msbar=False
+            is_msbar=False,
         )
         np.testing.assert_allclose(res_s, 0.0)
 
@@ -141,7 +141,7 @@ def test_quad_ker(monkeypatch):
             backward_method="expanded",
             a_s=0.0,
             L=0.0,
-            is_msbar=False
+            is_msbar=False,
         )
         if label[-1] == label[-2]:
             np.testing.assert_allclose(res_ns, 1.0)
@@ -172,7 +172,7 @@ def test_quad_ker(monkeypatch):
             backward_method="exact",
             a_s=0.0,
             L=0.0,
-            is_msbar=False
+            is_msbar=False,
         )
         if label[-1] == label[-2]:
             np.testing.assert_allclose(res_ns, 1.0)
@@ -190,7 +190,7 @@ def test_quad_ker(monkeypatch):
         backward_method=None,
         a_s=0.0,
         L=0.0,
-        is_msbar=False
+        is_msbar=False,
     )
     np.testing.assert_allclose(res_ns, 0.0)
